requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.10
requests==2.32.3